requirements.txt
cassandra-driver==3.0.0
flask==0.10.1